[pipelineX](minor) refine code
