+++ be/src/pipeline/pipeline_x/dependency.h
@@ -111,7 +111,7 @@ protected:
 class WriteDependency : public Dependency {
 public:
     WriteDependency(int id, std::string name) : Dependency(id, name), _ready_for_write(true) {}
-    virtual ~WriteDependency() = default;
+    ~WriteDependency() override = default;
 
     bool is_write_dependency() override { return true; }
 
@@ -428,7 +428,7 @@ private:
 
 struct MultiCastSharedState {
 public:
-    std::shared_ptr<pipeline::MultiCastDataStreamer> _multi_cast_data_streamer;
+    std::shared_ptr<pipeline::MultiCastDataStreamer> multi_cast_data_streamer;
 };
 
 class MultiCastDependency final : public WriteDependency {
@@ -438,7 +438,7 @@ public:
     ~MultiCastDependency() override = default;
     void* shared_state() override { return (void*)&_multi_cast_state; };
     MultiCastDependency* can_read(const int consumer_id) {
-        if (_multi_cast_state._multi_cast_data_streamer->can_read(consumer_id)) {
+        if (_multi_cast_state.multi_cast_data_streamer->can_read(consumer_id)) {
             return nullptr;
         } else {
             return this;
